made it so we can look and test coverage and made index and utils test files and tests
