--- backend/jest.config.js
@@ -1,18 +1,0 @@
-export default {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  extensionsToTreatAsEsm: [".ts"],
-  transform: {
-    "^.+\\.ts$": [
-      "ts-jest",
-      {
-        useESM: true,
-      },
-    ],
-  },
-  moduleFileExtensions: ["ts", "js"],
-  testMatch: ["**/*.test.ts", "**/*.test.js"],
-  moduleNameMapper: {
-    "^(\\.{1,2}/.*)\\.js$": "$1",
-  },
-};
