--- backend/tests/unit/utils.test.ts
@@ -1,14 +1,0 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import Utils from "../../src/utility/utils";
-
-describe("Testing utils", () => {
-  test("Testing isValidMySQLDate gives true for '2025-03-16'", () => {
-    const response = Utils.isValidMySQLDate("2025-03-16");
-    expect(response).toBe(true);
-  });
-
-  test("Testing isValidMySQLDate gives false for '03/16/2025'", () => {
-    const response = Utils.isValidMySQLDate("03/16/2025");
-    expect(response).toBe(false);
-  });
-});
